\name{alignment}
\alias{alignment}
\title{
Retrieves the alignment slot of the APPT.list object 
}
\description{
Retrieves the alignment slot of the object APPT.list, so that the user doesnt need to touch the object directly. 
}
\usage{
alignment(value)
}
\arguments{
\item{value}{
}
}

\value{
A matrix, created from a seqinr alignment object 
}
\author{
Lucia Spangenberg, Hugo Naya
}


