\name{bcool-package}
\alias{bcool-package}
\alias{bcool}
\docType{package}
\title{
bcool: Labels Used for Candidate Identification in Alignments
}
\description{
	The method presented here identifies associations between amino acid changes 
	(taking into account several amino acid properties) in interesting positions 
	in an alignment and some meta information (e. g. phenotypic data). 
	Even though the method has general applicability (other organisms,different 
	amino acid properties, different meta data), it was tested on a set of 209 
	bacteria of several genera and the goal was to find amino acid changes that 
	might be correlated with the pathogenicity of the bacteria.
}
\details{
\tabular{ll}{
Package: \tab bcool\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-30\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab ape, seqinr, MCMCglmm, methods, doMPI\cr
}

}
\author{
Lucia Spangenberg, Hugo Naya
Maintainer: Hugo Naya <naya@pasteur.edu.uy>

}
\references{
  L. Spangenberg, F. Battke, M. Grana, K. Nieselt and H. Naya (2011)
  Identifying associations between amino acid changes and meta information in alignments

}

\keyword{ package }

