\name{levelsMCMCglmm}
\alias{levelsMCMCglmm}
\title{
Retrieves the levels of the classes considered
}
\description{
Retrieves the levels of the classes (labels) that are considered. For example, if the label pathogenicity is considered, the method returns the level of it: yes and no.
}
\usage{
levelsMCMCglmm(value)
}
\arguments{
  \item{value}{
}
}

\value{

character vector, with the levels of the classes
}
\author{Lucia Spangenberg \email{lucia@pasteur.edu.uy} and Hugo Naya \email{naya@pasteur.edu.uy}}
