\name{multiMCMCglmm}
\alias{multiMCMCglmm}
\title{
Retrieves the multiMCMCglmm slot of the APPT.list object 
}
\description{
Retrieves the multiMCMCglmm slot of the object APPT.list, so that the user doesn0x00B4t need to access the object directly.
}
\usage{
multiMCMCglmm(value)
}
\arguments{
  \item{value}{
}
}
\value{

A list of MCMCglmm runs
}
\author{Lucia Spangenberg \email{lucia@pasteur.edu.uy} and Hugo Naya \email{naya@pasteur.edu.uy}}


