\name{phylo-class}
\Rdversion{1.1}
\docType{class}
\alias{phylo-class}

\title{Class "phylo"}
\description{
	S4 version of the phylo class in ape package
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("phylo", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "phylo" in the signature.
}
\keyword{classes}
